pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "recording-angel-api"
version = "0.1.0"
description = "Real-time audio transcription API with AI paragraph organization"
authors = [
    {name = "Recording Angel Team"}
]
keywords = ["transcription", "real-time", "websocket", "assemblyai", "gemini"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]==0.112.2",
    "uvicorn[standard]==0.30.6",
    "python-dotenv==1.0.1",
    "httpx==0.27.2",
    "websockets==12.0",
    "assemblyai==0.34.0",
    "google-generativeai==0.8.3",
    "google-cloud-translate==3.15.3",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "python-multipart==0.0.9",
    "sqlalchemy==2.0.43",
    "alembic==1.13.1",
    "email-validator==2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
    "ruff>=0.1.0",
]

[project.urls]
Homepage = "https://recordingangel.org"
Repository = "https://github.com/Pharoh-Ramses/Recording-Angel"
Issues = "https://github.com/Pharoh-Ramses/Recording-Angel/issues"

[tool.hatch.build.targets.wheel]
packages = ["app"]

# Tool configurations
[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.ruff]
target-version = "py311"
line-length = 100
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"
